update readme fix autocmd of zsh

+++ zsh/config/autocmd.zsh
@@ -1,7 +1,7 @@
 #!/bin/zsh
 
 # Start ssh-agent outside tmux in silent
-if [ -z $TMUX ]
+if [ -z "$TMUX" ] && [ -n "$PS1" ]
 then
     pkill ssh-agent
     eval "$(ssh-agent -s)" > /dev/null 2>&1
